+++ ci/tasks/promote-candidate.sh
@@ -5,15 +5,15 @@ set -e
 : ${ALICLOUD_ACCESS_KEY_ID:?}
 : ${ALICLOUD_ACCESS_KEY_SECRET:?}
 
+source bosh-cpi-src/ci/tasks/utils.sh
 source /etc/profile.d/chruby.sh
-chruby 2.1.2
+chruby 2.4.4
 
 # Creates an integer version number from the semantic version format
 # May be changed when we decide to fully use semantic versions for releases
 integer_version=`cut -d "." -f1 release-version-semver/number`
 echo $integer_version > promoted/integer_version
 
-# debug
 cp -r bosh-cpi-src promoted/repo
 
 dev_release=$(echo $PWD/bosh-cpi-release/*.tgz)
@@ -23,13 +23,10 @@ pushd promoted/repo
   cat > config/private.yml << EOF
 ---
 blobstore:
-  s3:
+  provider: s3
     access_key_id: $ALICLOUD_ACCESS_KEY_ID
     secret_access_key: $ALICLOUD_ACCESS_KEY_SECRET
 EOF
-
-  echo "using bosh CLI version..."
-  bosh version
 
   echo "finalizing CPI release..."
   bosh finalize release ${dev_release} --version $integer_version
@@ -39,7 +36,7 @@ EOF
   git diff | cat
   git add .
 
-  git config --global user.email demon.wy@alibaba-inc.com
-  git config --global user.name demonwy
+  git config --global user.email guimin.hgm@alibaba-inc.com
+  git config --global user.name xiaozhu36
   git commit -m "New final release v $integer_version"
 popd